Updates workspace context to object
Changes the `currentWorkspace` prop type from string to the `Workspace` object across several components.

This allows accessing workspace properties, such as name and slug, directly within the components, instead of only the name.

Specifically, it enables displaying the workspace name in the workspace switcher and creating a link to manage members for the currently selected workspace.