Refactors analytics view for breakdown data
Renames and refactors the AnalyticsView component to BreakdownAnalyticsView for clarity.

This change enhances the component's flexibility by allowing the specification of available views (graph, table, heatmap), and removing unnecessary props when they are not used in certain views.

The breakdown per cities view now utilizes the availableViews prop to allow a heatmap visualization as well as the table view.