Refactors RTS page layout for better structure
Introduces a layout component for the RTS pages to encapsulate the common header and navigation, promoting code reuse and maintainability.

Moves the shared header and navigation elements into a dedicated layout component and uses this layout in the "For Delivery Today" page, resulting in a cleaner and more organized component structure.

Also, adjusts the useEffect hook in the "For Delivery Today" page to prevent unnecessary re-renders when the page name search value hasn't actually changed.